fix(mcp): add blocklist + geocode-name guards to the NBCN climate resolver
resolveSmnStation and resolveForecastPoint both reject international city names
(isBlocklisted) and geocode hits that don't textually resemble the query
(geocodedNameMatchesQuery). resolveNbcnStation — behind meteoswissClimateData —
had neither and used a wider 80 km radius, so {station: "Paris"} geocoded the
hamlet Paris (VD) and returned Payerne's decades of climate data as a success.

- Add both guards to resolveNbcnStation, mirroring the SMN resolver.
- De-duplicate geocodedNameMatchesQuery (was copied identically in two data
  files) into name-matcher.ts so the guard cannot drift between resolvers again.

Addresses FUN-1 from the 2026-07-11 security review.
