+++ packages/meteoswiss-mcp/src/data/ogd-station-resolver.ts
@@ -11,7 +11,7 @@ import { debugData } from '../support/logging.js';
 import { OGD_COLLECTIONS } from '../schemas/ogd-shared.js';
 import type { ForecastPoint } from '../schemas/ogd-shared.js';
 import { normalize } from '../support/normalize.js';
-import { scoreNameMatch } from '../support/name-matcher.js';
+import { scoreNameMatch, geocodedNameMatchesQuery } from '../support/name-matcher.js';
 import { findNearest } from '../support/haversine.js';
 import { geocodeSwissLocation, type GeocodeOrigin } from '../support/geocode.js';
 import { classifyQuery } from '../support/query-classifier.js';
@@ -86,29 +86,6 @@ export type ResolveResult = {
   alternatives: ForecastPoint[];
   confidence: 'exact' | 'fuzzy';
 };
-
-/**
- * Returns true if the user's query has at least one token (≥3 chars) that
- * appears as a substring in the geocoded place name, or vice versa.
- * Guards against gibberish queries (e.g., "NOTASTATION") that the live
- * swisstopo API fuzzy-matches to an unrelated Swiss coordinate.
- */
-function geocodedNameMatchesQuery(query: string, geocodedName: string): boolean {
-  const queryNorm = normalize(query.trim());
-  const nameNorm = normalize(geocodedName);
-
-  const queryTokens = queryNorm.split(/\s+/).filter((t) => t.length >= 3);
-  for (const token of queryTokens) {
-    if (nameNorm.includes(token)) return true;
-  }
-
-  const nameTokens = nameNorm.split(/\s+/).filter((t) => t.length >= 3);
-  for (const token of nameTokens) {
-    if (queryNorm.includes(token)) return true;
-  }
-
-  return false;
-}
 
 /**
  * Resolve a query string to a forecast point.
